requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
statsmodels>=0.14.0
yfinance>=0.2.0
pydantic>=2.0.0
pyyaml>=6.0.0
tqdm>=4.65.0
pytest>=7.0.0
